Extract buildercore.context_handler.only_if
