+++ src/buildercore/context_handler.py
@@ -53,5 +53,17 @@ def download_from_s3(stackname, refresh=False):
     s3.download(key, expected_path)
     return True
 
+def only_if(servicename):
+    """Decorator that only executes an update function if the context contains a particular servicename that would need it"""
+    def wrap1(fn):
+        def wrap2(stackname, context, **kwargs):
+            # only update service if stack is using given service
+            LOG.info("Try to update '%s'", servicename)
+            if context.get(servicename):
+                return fn(stackname, context, **kwargs)
+            LOG.info("Skipped '%s' as not in the context", servicename)
+        return wrap2
+    return wrap1
+
 class MissingContextFile(RuntimeError):
     pass
