+++ src/buildercore/terraform.py
@@ -4,6 +4,7 @@ from os.path import join
 from python_terraform import Terraform
 from buildercore.utils import ensure
 from .config import BUILDER_BUCKET, BUILDER_REGION, TERRAFORM_DIR, ConfigurationError
+from .context_handler import only_if
 
 RESOURCE_TYPE_FASTLY = 'fastly_service_v1'
 RESOURCE_NAME_FASTLY = 'fastly-cdn'
@@ -58,7 +59,8 @@ def init(stackname):
     t.init(input=False, capture_output=False, raise_on_error=True)
     return t
 
-def update(stackname):
+@only_if('fastly')
+def update(stackname, context):
     ensure('FASTLY_API_KEY' in os.environ, "a FASTLY_API_KEY environment variable is required to provision Fastly resources", ConfigurationError)
     t = init(stackname)
     t.apply(input=False, capture_output=False, raise_on_error=True)
